render: Refactored to support dynamic render target setup.
Now, we can switch the render target of a renderer dynamically. With this, we
can create one EGLContext for multiple native windows so that we can save OpenGL
resources and texture download cost.

Additionally, some renderer APIs are added for surface attach and damage flush.

Modules that were using old APIs are modified to use new APIs. Essentially, the
root object (i.e. pepper_drm_t) creates renderers, and each output chooses its
desired renderer and points to the renderer. pepper_render_target_t is created
for an output and it is set as a render target of the renderer when repaint.

Some messy codes are cleaned up while rewriting for the new APIs. But still got
a lot more, especially, codes taken from weston.

Change-Id: I5c9796c44f74efa3aa735396f58c6499a817937b