pepper: Accumulate output damage and automatically schedule repaint.
Damage is maintained per output and is passed to the repaint function.
Damage region is in output local coordinate space so an output backend should
be careful not to translate it to global compositor space.

No damage, we have nothing to repaint. So, we can safely schedule repaint
whenever damage is generated. Essentially, we don't have to expose repaint
scheduling APIs.

Change-Id: Iebff5fec5f93d764a8ea8894b7f84c8ded9c59c3