pepper: Attach surface to outputs on surface.commit().
Attaching a surface is a process of notifying output backends so that they can
prepare for reading it and allocate resources for it.

For some types of buffers like EGL, the size is not known until it is queried
to the native library. although we can get the size of a shm buffer using
wl_shm, we basically assume that the buffer size should be queried to output
backends. If the backend uses pepper_render, then it can simply redirect the
request to the renderer.

Change-Id: Ie145f9029447605116a81984431cdc5235dc1eca