pepper: Better ID allocation algorithm
Avoid allocating an ID having same value with recently released IDs.
To do this, we use some upper bits of the ID as an age. When an ID
is reused, the age is increased. This can help detect error cases
where users incorrectly access IDs that has been already destroyed.

But, this does not mean that we are completely safe from the problem.
Some kind of locking mechanism or reference counting is required to
prevent from accessing invalid (most likely dead) IDs.

Change-Id: I218f94566c0f2e63d23df4e8fa1a2890184ca583